Specification says rand() returns a signed number. Protect against this.

+++ model.c
@@ -30,7 +30,7 @@ void cf_grid_init( cf_grid_t *const this_grid )
 
                         while ( --col >= 0 )
                         {
-                                cf_cellState_t state = rand() % CF_COLORCOUNT; // not exactly uniform but anyway
+                                cf_cellState_t state = ( ( uint8_t )rand() ) % CF_COLORCOUNT; // not exactly uniform but anyway
                                 this_grid->cell[row][col] = state;
                                 //printf( " c=%d", col );
                         }
